Add files via upload
The Insecure Juice Shop project is a vulnerable web application that is intentionally designed to have security flaws. It is used as a training tool for security professionals to practice identifying and exploiting vulnerabilities in a safe and controlled environment. The project is built using modern web technologies and includes a variety of vulnerabilities, including injection attacks, cross-site scripting (XSS), and broken authentication and session management. By working with the Insecure Juice Shop